requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.10
idna==3.3
lxml==4.7.1
mysql-connector-python==8.0.28
protobuf==3.19.3
pyTelegramBotAPI==4.3.1
python-telegram-bot-pagination==0.0.2
requests==2.27.1
soupsieve==2.3.1
urllib3==1.26.8
